GPU/requirements.txt
torch
diffusers
transformers
accelerate
flask
grpcio
grpcio-tools
protobuf
streamlit
pillow
